fix: fix misstakes from test

+++ internal/domain/models/User.go
@@ -1,7 +1,7 @@
 package models
 
 type User struct {
-	Id           int32  `json:"id" db:"id"`
+	Id           int64  `json:"id" db:"id"`
 	Name         string `json:"name" db:"name"`
 	Email        string `json:"email" db:"email"`
 	PasswordHash []byte `json:"password_hash" db:"password_hash"`
